requirements.txt
Flask
transformers
torch
flask-cors
sentencepiece
openai-whisper